+++ internal/assets-service/repositories/cosmetics/repository.go
@@ -21,11 +21,11 @@ type CosmeticsRepository interface {
 
 	GetCosmeticsListByWorld(worldId uuid.UUID, offset int, limit int) ([]*models.Cosmetic, int64, error)
 
-	CreateCosmetic(category uuid.UUID, worldId uuid.UUID, price float64, cosmetic *models.Cosmetic, userId uuid.UUID) error
+	CreateCosmetic(category uuid.UUID, worldId uuid.UUID, price int64, cosmetic *models.Cosmetic, userId uuid.UUID) error
 
 	GetCosmeticByUrlCategoryAndWorld(url string, categoryId uuid.UUID, worldId uuid.UUID) (*models.Cosmetic, error)
 
-	UpdateCosmetic(cosmeticId uuid.UUID, price float64, url string) error
+	UpdateCosmetic(cosmeticId uuid.UUID, price int64, url string) error
 
 	DeleteCosmetic(cosmeticId uuid.UUID) error
 }
